move readResponse and writeLibrarianState to state.go
