+++ internal/docker/docker.go
@@ -223,7 +223,7 @@ func (c *Docker) Configure(ctx context.Context, request *ConfigureRequest) (stri
 	}
 
 	// Read the new library state from the response.
-	libraryState, err := readResponse(
+	libraryState, err := config.ReadResponse(
 		func(data []byte, libraryState *config.LibraryState) error {
 			return json.Unmarshal(data, libraryState)
 		},
@@ -241,7 +241,7 @@ func (c *Docker) Configure(ctx context.Context, request *ConfigureRequest) (stri
 	}
 
 	// Write the updated librarian state to state.yaml.
-	if err := writeLibrarianState(
+	if err := config.WriteLibrarianState(
 		func(state *config.LibrarianState) ([]byte, error) {
 			return yaml.Marshal(state)
 		},
@@ -330,50 +330,6 @@ func writeRequest(state *config.LibrarianState, libraryID, jsonFilePath string) 
 		if err != nil {
 			return fmt.Errorf("failed to write generate request JSON file: %w", err)
 		}
-	}
-
-	return nil
-}
-
-// readResponse reads the library state from configure-response.json.
-//
-// The response file is removed afterwards.
-func readResponse(contentLoader func(data []byte, state *config.LibraryState) error, jsonFilePath string) (*config.LibraryState, error) {
-	data, err := os.ReadFile(jsonFilePath)
-	defer func(name string) {
-		err := os.Remove(name)
-		if err != nil {
-			slog.Warn("fail to remove file", slog.String("name", name), slog.Any("err", err))
-		}
-	}(jsonFilePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read response file, path: %s, error: %w", jsonFilePath, err)
-	}
-
-	libraryState := &config.LibraryState{}
-
-	if err := contentLoader(data, libraryState); err != nil {
-		return nil, fmt.Errorf("failed to load file, %s, to state: %w", jsonFilePath, err)
-	}
-
-	return libraryState, nil
-}
-
-// writeLibrarianState writes the given librarian state to a yaml file.
-func writeLibrarianState(contentParser func(state *config.LibrarianState) ([]byte, error), state *config.LibrarianState, yamlFilePath string) error {
-	yamlFile, err := os.Create(yamlFilePath)
-	if err != nil {
-		return fmt.Errorf("failed to create librarian state file, path: %s, error: %w", yamlFilePath, err)
-	}
-	defer yamlFile.Close()
-
-	data, err := contentParser(state)
-	if err != nil {
-		return fmt.Errorf("failed to convert state to bytes: %w", err)
-	}
-
-	if _, err := yamlFile.Write(data); err != nil {
-		return fmt.Errorf("failed to write librarian state file, path: %s, error: %w", yamlFilePath, err)
 	}
 
 	return nil
